% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carregar.R
\name{carregar_textos}
\alias{carregar_textos}
\title{Carregar dados}
\usage{
carregar_textos(x)
}
\arguments{
\item{x}{vetor de chr com textos que deseja carregar}
}
\value{
corpus com os textos
}
\description{
Carregar dados
}

