% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{encontrar_erros}
\alias{encontrar_erros}
\title{Encontrar erros}
\usage{
encontrar_erros(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor com as palavras que estão escritas erradas o texto.
}
\description{
Encontrar erros
}

