% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hunspell.R
\name{look_SFX}
\alias{look_SFX}
\title{Procurar os sufixos em um dicionário}
\usage{
look_SFX(aff)
}
\arguments{
\item{aff}{dicionário de sufixos e afixos no estilo do do hunspell.}
}
\description{
Procurar os sufixos em um dicionário
}

