% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelos.R
\name{mtd}
\alias{mtd}
\title{Matriz de termos e documentos}
\usage{
mtd(s, tipo = "contagem")
}
\arguments{
\item{s}{vetor de chr}

\item{tipo}{tipo de matriz (contagem ou flag)}
}
\value{
uma matriz de termos e documentos
}
\description{
Matriz de termos e documentos
}

