% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelos.R
\name{ngram}
\alias{ngram}
\title{NGRAM}
\usage{
ngram(s, n = 1)
}
\arguments{
\item{s}{vetor de chr}

\item{n}{n do n-gram (quantas palavras juntas)}
}
\value{
uma lista com os n-grams para cada elemento do vetor.
}
\description{
NGRAM
}
\seealso{
\code{\link{mtd}}
}

