% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelos.R
\name{palavras_seguidas}
\alias{palavras_seguidas}
\title{Palavras seguidas}
\usage{
palavras_seguidas(s, p, qtd = NULL)
}
\arguments{
\item{s}{vetor de chr}

\item{p}{palavra que deseja buscar}

\item{qtd}{quantidade de palavras que você quer mostrar (se = NULL, mostrar 
todos)}
}
\value{
um data.frame com as colunas ng_a palavra que mais veio antes e n_a 
a quantidade de vezes (qtd). ng_d são as palavras que aparecem depois do termo p.
}
\description{
Quais são as palavras que mais aparecem antes ou depois da palavra p.
}

