% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remove_pontuacao}
\alias{remove_pontuacao}
\title{Remover pontuação}
\usage{
remove_pontuacao(x)
}
\arguments{
\item{vetor}{de chr}
}
\value{
vetor de chr com a pontuação removida
}
\description{
Remover pontuação
}

