% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remover_acentos}
\alias{remover_acentos}
\title{Remover acentos}
\usage{
remover_acentos(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr com os acentos removidos
}
\description{
Remover acentos
}
\seealso{
Other transf: \code{\link{remover_dinheiro}},
  \code{\link{remover_espacos_excedentes}},
  \code{\link{remover_hashtags}},
  \code{\link{remover_numeros}},
  \code{\link{remover_palavras}},
  \code{\link{remover_palavra}},
  \code{\link{remover_pontuacao}},
  \code{\link{remover_stopwords}},
  \code{\link{remover_url}},
  \code{\link{substituir_palavras}},
  \code{\link{substituir_palavra}},
  \code{\link{transformar_minusculo}}
}

