% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remover_dinheiro}
\alias{remover_dinheiro}
\title{Remover dinheiro}
\usage{
remover_dinheiro(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr com r$ excluido
}
\description{
Remover dinheiro
}
\note{
deve ser usada depois de transformar o texto em minúsculo.
}
\seealso{
Other transf: \code{\link{remover_acentos}},
  \code{\link{remover_espacos_excedentes}},
  \code{\link{remover_hashtags}},
  \code{\link{remover_numeros}},
  \code{\link{remover_palavras}},
  \code{\link{remover_palavra}},
  \code{\link{remover_pontuacao}},
  \code{\link{remover_stopwords}},
  \code{\link{remover_url}},
  \code{\link{substituir_palavras}},
  \code{\link{substituir_palavra}},
  \code{\link{transformar_minusculo}}
}

