% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remover_espacos_excedentes}
\alias{remover_espacos_excedentes}
\title{Remover espaços excedentes}
\usage{
remover_espacos_excedentes(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr seme spaços excedentes
}
\description{
Remover espaços excedentes
}

