% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remover_hashtags}
\alias{remover_hashtags}
\title{Remover hashtags}
\usage{
remover_hashtags(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr com hashtags excluidas
}
\description{
Remover hashtags
}

