% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remover_numeros}
\alias{remover_numeros}
\title{Remover números}
\usage{
remover_numeros(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr sem números
}
\description{
Remover números
}
\seealso{
Other transf: \code{\link{remover_acentos}},
  \code{\link{remover_dinheiro}},
  \code{\link{remover_espacos_excedentes}},
  \code{\link{remover_hashtags}},
  \code{\link{remover_palavras_erradas}},
  \code{\link{remover_palavras}},
  \code{\link{remover_palavra}},
  \code{\link{remover_pontuacao}},
  \code{\link{remover_stopwords}},
  \code{\link{remover_url}},
  \code{\link{substituir_palavras}},
  \code{\link{substituir_palavra}},
  \code{\link{transformar_corrigir}},
  \code{\link{transformar_minusculo}}
}

