% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remover_palavra}
\alias{remover_palavra}
\title{Remover palavra}
\usage{
remover_palavra(s, p)
}
\arguments{
\item{s}{vetor de chr}

\item{p}{palavra que deseja excluir}
}
\value{
vetor de chr sem a palavra p
}
\description{
Remover palavra
}
\seealso{
Other transf: \code{\link{remover_acentos}},
  \code{\link{remover_dinheiro}},
  \code{\link{remover_espacos_excedentes}},
  \code{\link{remover_hashtags}},
  \code{\link{remover_numeros}},
  \code{\link{remover_palavras}},
  \code{\link{remover_pontuacao}},
  \code{\link{remover_stopwords}},
  \code{\link{remover_url}},
  \code{\link{substituir_palavras}},
  \code{\link{substituir_palavra}},
  \code{\link{transformar_minusculo}}
}

