% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remover_palavras}
\alias{remover_palavras}
\title{Remover palavras}
\usage{
remover_palavras(s, p)
}
\arguments{
\item{s}{vetor de chr}

\item{p}{vetor de palavras que desejar excluir}
}
\value{
vetor de chr sem as palavras indicadas no parametro p
}
\description{
Remover palavras
}

