% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remover_palavras_erradas}
\alias{remover_palavras_erradas}
\title{Remover palavras erradas}
\usage{
remover_palavras_erradas(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr com as palavras escritas de forma errada removidas.
}
\description{
Remover palavras erradas
}

