% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remover_pontuacao}
\alias{remover_pontuacao}
\title{Remover pontuação}
\usage{
remover_pontuacao(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr com a pontuação removida
}
\description{
Remover pontuação
}

