% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remover_stopwords}
\alias{remover_stopwords}
\title{Remover Stopbords}
\usage{
remover_stopwords(s, ignorar = character(0))
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr sem as stopwords
}
\description{
Remover Stopbords
}
\seealso{
\code{\link{stopwords}}
}

