% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remover_url}
\alias{remover_url}
\title{Remover urls}
\usage{
remover_url(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr com urls excluidas
}
\description{
Remover urls
}
\seealso{
Other transf: \code{\link{remover_acentos}},
  \code{\link{remover_dinheiro}},
  \code{\link{remover_espacos_excedentes}},
  \code{\link{remover_hashtags}},
  \code{\link{remover_numeros}},
  \code{\link{remover_palavras}},
  \code{\link{remover_palavra}},
  \code{\link{remover_pontuacao}},
  \code{\link{remover_stopwords}},
  \code{\link{substituir_palavras}},
  \code{\link{substituir_palavra}},
  \code{\link{transformar_minusculo}}
}

