% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{stopwords}
\alias{stopwords}
\title{Stopwords}
\usage{
stopwords()
}
\value{
vetor com palavras consideradas stopwords
}
\description{
Palavras que geralemente não trazem informações relevantes para análise dos 
textos.
}

