% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemmer.R, R/transformacoes.R
\name{substituir_palavra}
\alias{substituir_palavra}
\title{Substituir uma palavra por outra}
\usage{
substituir_palavra(s, pa, pd)

substituir_palavra(s, pa, pd)
}
\arguments{
\item{s}{vetor de chr}

\item{pa}{palavra que deve ser substituida}

\item{pd}{palavra colocada no lugar de pa}

\item{s}{vetor de chr}

\item{pa}{palavra que deve ser substituida}

\item{pd}{palavra colocada no lugar de pa}
}
\value{
vetor de chr com as palavras substituidas

vetor de chr com as palavras substituidas
}
\description{
Substituir uma palavra por outra

Substituir uma palavra por outra
}

