% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{transf_minusculo}
\alias{transf_minusculo}
\title{Tranformar letras do corpus em minusculo}
\usage{
transf_minusculo(x)
}
\arguments{
\item{vetor}{de chr}
}
\value{
vetor de chr transformado em minúsculos
}
\description{
Tranformar letras do corpus em minusculo
}

