% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{transformacoes}
\alias{transformacoes}
\title{Tranformações}
\description{
Lista de todas as transformações que podem ser feitas com o pacote
}
\seealso{
\code{\link{transformar_minusculo}}

\code{\link{remover_pontuacao}}

\code{\link{remover_url}}

\code{\link{remover_url}}

\code{\link{remover_hashtags}}

\code{\link{remover_dinheiro}}

\code{\link{remover_numeros}}

\code{\link{remover_espacos_excedentes}}

\code{\link{remover_acentos}}

\code{\link{remover_stopwords}}

\code{\link{transformar_stemming}}

\code{\link{transformar_corrigir}}

\code{\link{remover_palavras_erradas}}
}

