% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{transformar_corrigir}
\alias{transformar_corrigir}
\title{Corrigir as palavras}
\usage{
transformar_corrigir(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr com palavras corrigidas usando o hunspell
}
\description{
Corrigir as palavras
}
\seealso{
\code{\link[hunspell]{hunspell_check_pt}} and 
\code{\link[hunspell]{hunspell_suggest_pt}}
}

