% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{transformar_minusculo}
\alias{transformar_minusculo}
\title{Tranformar letras do corpus em minusculo}
\usage{
transformar_minusculo(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr transformado em minúsculos
}
\description{
Tranformar letras do corpus em minusculo
}

