% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{transformar_stemming}
\alias{transformar_stemming}
\title{Fazer o Stemming}
\usage{
transformar_stemming(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr com stemming realizado
}
\description{
Fazer o Stemming
}

