% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remover_acentos}
\alias{remover_acentos}
\title{Remover acentos}
\usage{
remover_acentos(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr com os acentos removidos
}
\description{
Remover acentos
}

