% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remover_dinheiro}
\alias{remover_dinheiro}
\title{Remover dinheiro}
\usage{
remover_dinheiro(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr com r$ excluido
}
\description{
Remover dinheiro
}
\note{
deve ser usada depois de transformar o texto em minúsculo.
}

