% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remover_numeros}
\alias{remover_numeros}
\title{Remover números}
\usage{
remover_numeros(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr sem números
}
\description{
Remover números
}

