% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remover_palavra}
\alias{remover_palavra}
\title{Remover palavra}
\usage{
remover_palavra(s, p)
}
\arguments{
\item{s}{vetor de chr}

\item{p}{palavra que deseja excluir}
}
\value{
vetor de chr sem a palavra p
}
\description{
Remover palavra
}

