% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{remover_palavras}
\alias{remover_palavras}
\title{Remover palavras}
\usage{
remover_palavras(s, p)
}
\arguments{
\item{s}{vetor de chr}

\item{p}{vetor de palavras que desejar excluir}
}
\value{
vetor de chr sem as palavras indicadas no parametro p
}
\description{
Remover palavras
}
\seealso{
Other transf: \code{\link{remover_acentos}},
  \code{\link{remover_dinheiro}},
  \code{\link{remover_espacos_excedentes}},
  \code{\link{remover_hashtags}},
  \code{\link{remover_numeros}},
  \code{\link{remover_palavra}},
  \code{\link{remover_pontuacao}},
  \code{\link{remover_stopwords}},
  \code{\link{remover_url}},
  \code{\link{substituir_palavras}},
  \code{\link{substituir_palavra}},
  \code{\link{transformar_minusculo}}
}

