% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{substituir_palavra}
\alias{substituir_palavra}
\title{Substituir uma palavra por outra}
\usage{
substituir_palavra(s, pa, pd)
}
\arguments{
\item{s}{vetor de chr}

\item{pa}{palavra que deve ser substituida}

\item{pd}{palavra colocada no lugar de pa}
}
\value{
vetor de chr com as palavras substituidas
}
\description{
Substituir uma palavra por outra
}
\seealso{
Other transf: \code{\link{remover_acentos}},
  \code{\link{remover_dinheiro}},
  \code{\link{remover_espacos_excedentes}},
  \code{\link{remover_hashtags}},
  \code{\link{remover_numeros}},
  \code{\link{remover_palavras}},
  \code{\link{remover_palavra}},
  \code{\link{remover_pontuacao}},
  \code{\link{remover_stopwords}},
  \code{\link{remover_url}},
  \code{\link{substituir_palavras}},
  \code{\link{transformar_minusculo}}
}

