% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemmer.R, R/transformacoes.R
\name{substituir_palavras}
\alias{substituir_palavras}
\title{Substituir uma lista de palavras por uma lista de palavras}
\usage{
substituir_palavras(s, pa, pd)

substituir_palavras(s, pa, pd)
}
\arguments{
\item{s}{vetor de chr}

\item{pa}{vetor de palavras que ser達o substituidas}

\item{pd}{vetor de palavras que ser達o colcoadas no lugar}

\item{s}{vetor de chr}

\item{pa}{vetor de palavras que ser達o substituidas}

\item{pd}{vetor de palavras que ser達o colcoadas no lugar}
}
\value{
vetor de chr com as palavras substituidas

vetor de chr com as palavras substituidas
}
\description{
Substituir uma lista de palavras por uma lista de palavras

Substituir uma lista de palavras por uma lista de palavras
}

