% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{substituir_palavras}
\alias{substituir_palavras}
\title{Substituir uma lista de palavras por uma lista de palavras}
\usage{
substituir_palavras(s, pa, pd)
}
\arguments{
\item{s}{vetor de chr}

\item{pa}{vetor de palavras que serão substituidas}

\item{pd}{vetor de palavras que serão colcoadas no lugar}
}
\value{
vetor de chr com as palavras substituidas
}
\description{
Substituir uma lista de palavras por uma lista de palavras
}
\seealso{
Other transf: \code{\link{remover_acentos}},
  \code{\link{remover_dinheiro}},
  \code{\link{remover_espacos_excedentes}},
  \code{\link{remover_hashtags}},
  \code{\link{remover_numeros}},
  \code{\link{remover_palavras_erradas}},
  \code{\link{remover_palavras}},
  \code{\link{remover_palavra}},
  \code{\link{remover_pontuacao}},
  \code{\link{remover_stopwords}},
  \code{\link{remover_url}},
  \code{\link{substituir_palavra}},
  \code{\link{transformar_corrigir}},
  \code{\link{transformar_minusculo}}
}

