% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corretor.R
\name{transformar_corrigir}
\alias{transformar_corrigir}
\title{Tranformar Corrigir}
\usage{
transformar_corrigir(s, dict = system.file("dic/Portuguese (Brazilian).dic",
  package = "PTtextmining"))
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor com as palavras corrigidas
}
\description{
Tranformar Corrigir
}
\note{
o Hunspell devolve um número de sugestões de correções. O código atual usa a sugestão que mais apareceu no
vetor e em caso de empate usa a ordem do Hunspell. Não sei se esse é o melhor jeito.
}
\seealso{
\code{\link[hunspell]{hunspell_check}} and 
\code{\link[hunspell]{hunspell_suggest}}
}

