% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{transformar_corrigir}
\alias{transformar_corrigir}
\title{Corrigir as palavras}
\usage{
transformar_corrigir(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr com palavras corrigidas usando o hunspell
}
\description{
Corrigir as palavras
}
\seealso{
\code{\link[hunspell]{hunspell_check_pt}} and 
\code{\link[hunspell]{hunspell_suggest_pt}}

Other transf: \code{\link{remover_acentos}},
  \code{\link{remover_dinheiro}},
  \code{\link{remover_espacos_excedentes}},
  \code{\link{remover_hashtags}},
  \code{\link{remover_numeros}},
  \code{\link{remover_palavras_erradas}},
  \code{\link{remover_palavras}},
  \code{\link{remover_palavra}},
  \code{\link{remover_pontuacao}},
  \code{\link{remover_stopwords}},
  \code{\link{remover_url}},
  \code{\link{substituir_palavras}},
  \code{\link{substituir_palavra}},
  \code{\link{transformar_minusculo}}
}

