% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corretor.R, R/transformacoes.R
\name{transformar_corrigir}
\alias{transformar_corrigir}
\title{Tranformar Corrigir}
\usage{
transformar_corrigir(s)

transformar_corrigir(s)
}
\arguments{
\item{s}{vetor de chr}

\item{s}{vetor de chr}
}
\value{
vetor de chr com palavras corrigidas usando o hunspell
}
\description{
Usa o Hunspell p/ corrigir o texto.

Corrigir as palavras
}
\note{
o Hunspell devolve um número de sugestões de correções. O código atual usa a sugestão que mais apareceu no
vetor e em caso de empate usa a ordem do Hunspell. Não sei se esse é o melhor jeito.
}
\seealso{
\code{\link[hunspell]{hunspell_check_pt}} and 
\code{\link[hunspell]{hunspell_suggest_pt}}

Other transf: \code{\link{remover_acentos}},
  \code{\link{remover_dinheiro}},
  \code{\link{remover_espacos_excedentes}},
  \code{\link{remover_hashtags}},
  \code{\link{remover_numeros}},
  \code{\link{remover_palavras_erradas}},
  \code{\link{remover_palavras}},
  \code{\link{remover_palavra}},
  \code{\link{remover_pontuacao}},
  \code{\link{remover_stopwords}},
  \code{\link{remover_url}},
  \code{\link{substituir_palavras}},
  \code{\link{substituir_palavra}},
  \code{\link{transformar_minusculo}}
}

