% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformacoes.R
\name{transformar_minusculo}
\alias{transformar_minusculo}
\title{Tranformar letras do corpus em minusculo}
\usage{
transformar_minusculo(s)
}
\arguments{
\item{s}{vetor de chr}
}
\value{
vetor de chr transformado em minúsculos
}
\description{
Tranformar letras do corpus em minusculo
}
\seealso{
Other transf: \code{\link{remover_acentos}},
  \code{\link{remover_dinheiro}},
  \code{\link{remover_espacos_excedentes}},
  \code{\link{remover_hashtags}},
  \code{\link{remover_numeros}},
  \code{\link{remover_palavras_erradas}},
  \code{\link{remover_palavras}},
  \code{\link{remover_palavra}},
  \code{\link{remover_pontuacao}},
  \code{\link{remover_stopwords}},
  \code{\link{remover_url}},
  \code{\link{substituir_palavras}},
  \code{\link{substituir_palavra}},
  \code{\link{transformar_corrigir}}
}

