% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemmer.R, R/transformacoes.R
\name{transformar_stemming}
\alias{transformar_stemming}
\title{Fazer o Stemming}
\usage{
transformar_stemming(s)

transformar_stemming(s)
}
\arguments{
\item{s}{vetor de chr}

\item{s}{vetor de chr}
}
\value{
vetor de chr com stemming realizado

vetor de chr com stemming realizado
}
\description{
Essa função usa a função stemDocument do pacote tm para realizar o stemming.
Por sua vez, o tm utiliza o algoritmo Porter para realizar o stemming. Esse 
algoritmo é descrito no seguinte link: 
http://snowball.tartarus.org/algorithms/portuguese/stemmer.html

Essa função usa a função stemDocument do pacote tm para realizar o stemming.
Por sua vez, o tm utiliza o algoritmo Porter para realizar o stemming. Esse 
algoritmo é descrito no seguinte link: 
http://snowball.tartarus.org/algorithms/portuguese/stemmer.html
}

